backend/requirements.txt
django==5
django-rest-framework
psycopg2-binary
django-ckeditor-5
django-admin-ordering==0.16.1
channels
channels_redis
six
django_cleanup
django-rest-knox~=5.0.2
django-storages
gunicorn
djangorestframework~=3.15.2